Добавить LikeService и вынести из EventService всю логику работы с лайками/дизлайками.
